Book Service Alter, DTO created

+++ src/main/java/com/baylor/se/lms/data/AdminRepository.java
@@ -18,7 +18,7 @@ import java.util.Optional;
 
 public interface AdminRepository extends UserBaseRepository<Admin>, CrudRepository<Admin, Long> {
     @Query("select  A from Admin A where A.id = ?1 and A.deleteFlag = false")
-    Optional<Admin> findAdminById(double id);
+    Optional<Admin> findAdminById(long id);
 
     Optional <List<Admin>> findAllByDeleteFlagFalse();
 }
